Add basic navigation for purchase split viewing and move history to group page
