+++ src/routes/groups/index.tsx
@@ -2,6 +2,7 @@ import { Settlement } from "@/routes/settlement";
 import GroupPage from "./group-id";
 import NewGroup from "./new-group";
 import { Purchase } from "@/routes/purchase";
+import { PurchaseDetails } from "@/components/purchase/PurchaseDetails";
 
 const groupsRoutes = {
   path: "groups",
@@ -17,6 +18,10 @@ const groupsRoutes = {
     {
       path: ":id/purchase",
       element: <Purchase />,
+    },
+    {
+      path: ":id/purchase/:purchaseId",
+      element: <PurchaseDetails />,
     },
     {
       path: ":id/settlement",
